src/lambda_codebase/initial_commit/bootstrap_repository/adf-build/shared/requirements.txt
# Install libs here that you might want in AWS CodeBuild
pytest==3.0.7
mock==2.0.0
boto3==1.9.89
pyyaml>=5.1
schema==0.7.1
jsii~=0.21.1
aws_cdk.aws_codepipeline==1.20
aws_cdk.aws_lambda_event_sources==1.20
aws-cdk.aws-codepipeline-actions==1.20
aws_cdk.aws_codebuild==1.20
aws_cdk.aws_codecommit==1.20
aws-cdk.aws-s3==1.20
aws-cdk-assets==1.20
aws-cdk.aws-iam==1.20
aws-cdk.aws-kms==1.20
aws-cdk.aws-lambda==1.20
aws-cdk.aws-sns==1.20
aws-cdk.aws_secretsmanager==1.20
aws-cdk.aws_ssm==1.20
aws-cdk.aws_events==1.20
aws_cdk.aws_cloudformation==1.20
aws_cdk.aws_certificatemanager==1.20
aws_cdk.aws_sqs==1.20
aws_cdk.aws_ec2==1.20
aws_cdk.aws_sns==1.20
aws_cdk.aws_elasticloadbalancing==1.20
aws_cdk.aws_s3_notifications==1.20
aws_cdk.aws_codedeploy==1.20
aws_cdk.aws_ecr==1.20
aws_cdk.aws_stepfunctions==1.20
aws_cdk.aws_servicediscovery==1.20
aws_cdk.aws_autoscaling_hooktargets==1.20
aws_cdk.aws_cloudfront==1.20
aws_cdk.aws_route53_targets==1.20
aws_cdk.aws_route53==1.20
aws_cdk.aws_ecs==1.20
aws_cdk.aws_events_targets==1.20
aws_cdk.aws_dynamodb==1.20
aws_cdk.aws_applicationautoscaling==1.20
aws_cdk.aws_autoscaling==1.20
aws_cdk.aws_autoscaling_common==1.20
aws_cdk.aws_s3_assets==1.20
aws_cdk.aws_ecr_assets==1.20
aws_cdk.aws_elasticloadbalancingv2==1.20
aws_cdk.aws_apigateway==1.20
aws_cdk.aws_logs==1.20
aws_cdk.aws_ecr==1.20
aws_cdk.aws_cloudwatch==1.20
aws_cdk.aws_kinesis==1.20
aws_cdk.aws_s3_assets==1.20
aws_cdk.aws_sns_subscriptions==1.20
aws-cdk.core==1.20